change maximumSize to double

+++ PVFS/VFS/VFS/Models/DiskProperties.cs
@@ -6,7 +6,7 @@ namespace VFS.VFS.Models {
         public string Name { get; set; }
         public int NumberOfBlocks { get; set; }
         public int NumberOfUsedBlocks { get; set; }
-        public int MaximumSize { get; set; }
+        public double MaximumSize { get; set; }
         public int BlockSize { get; set; }
 
         public static DiskProperties Load(BinaryReader reader) {
